+++ src/config/image_format.rs
@@ -11,6 +11,26 @@ pub enum ImageFormat {
     Png,
     /// TIFF image format.
     Tiff,
+    /// GIF image format.
+    Gif,
+    /// TGA image format.
+    Tga,
+    /// DDS image format.
+    Dds,
+    /// BMP image format.
+    Bmp,
+    /// ICO image format.
+    Ico,
+    /// Radiance HDR image format.
+    Hdr,
+    /// OpenEXR image format.
+    OpenExr,
+    /// General PNM image format.
+    Pnm,
+    /// Farbfeld image format.
+    Farbfeld,
+    /// QOI image format.
+    Qoi,
     /// JPEG XL image format.
     #[cfg(feature = "jxl")]
     JpegXl,
@@ -55,15 +75,25 @@ impl ImageFormat {
                 .to_lowercase()
                 .as_str()
             {
-                "jpg" | "jpeg" => Self::Jpeg,
-                "png" => Self::Png,
-                "tiff" | "tif" => Self::Tiff,
-                #[cfg(feature = "jxl")]
-                "jxl" => Self::JpegXl,
-                #[cfg(feature = "webp")]
-                "webp" => Self::WebP,
                 #[cfg(feature = "avif")]
                 "avif" => Self::Avif,
+                "jpg" | "jpeg" => Self::Jpeg,
+                "png" => Self::Png,
+                "gif" => Self::Gif,
+                #[cfg(feature = "webp")]
+                "webp" => Self::WebP,
+                "tif" | "tiff" => Self::Tiff,
+                "tga" => Self::Tga,
+                "dds" => Self::Dds,
+                "bmp" => Self::Bmp,
+                "ico" => Self::Ico,
+                "hdr" => Self::Hdr,
+                "exr" => Self::OpenExr,
+                "pbm" | "pam" | "ppm" | "pgm" => Self::Pnm,
+                "ff" | "farbfeld" => Self::Farbfeld,
+                "qoi" => Self::Qoi,
+                #[cfg(feature = "jxl")]
+                "jxl" => Self::JpegXl,
                 ext => return Err(ImageFormatError::Unknown(ext.to_string())),
             },
         )
@@ -107,11 +137,11 @@ mod tests {
 
     #[test]
     fn to_unknown() {
-        let img_format = ImageFormat::from_ext("bmp");
+        let img_format = ImageFormat::from_ext("qex");
         assert!(img_format.is_err());
         assert_eq!(
             img_format.unwrap_err(),
-            ImageFormatError::Unknown("bmp".to_string())
+            ImageFormatError::Unknown("qex".to_string())
         );
     }