+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 All notable changes to the Rimage library will be documented in this file.
 
-## v0.10.2
+## v0.10.3
+
+### Changes
+
+- Added support for all image formats that supported by `image` crate
+
+## [v0.10.2](https://github.com/SalOne22/rimage/releases/tag/v0.10.2)
 
 ### Bug Fixes
 
@@ -95,7 +101,7 @@ All notable changes to the Rimage library will be documented in this file.
 
 ### Enhancements
 
-- Updated progress bar. ![Updated version](./assets/progress_bar.gif)
+- Updated progress bar. ![progress_bar](https://github.com/SalOne22/rimage/assets/111443297/847d30df-54e4-40c8-9d02-f67a66f140a8)
 - Rimage now uses the rayon crate for parallel optimizations.
 - Added `--quiet` flag that disables the progress bar.